requirements.txt
flask==2.3.3
langchain==0.1.16
langchain-openai==0.1.7
langchain-pinecone==0.1.0
pinecone-client==3.2.2
sentence-transformers==2.2.2
huggingface-hub==0.17.3
transformers==4.33.2
torch==2.0.1
ctransformers==0.2.27
pypdf==3.15.2
python-dotenv==1.0.0
setuptools==67.6.1
tqdm==4.64.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
-e .